+++ admin/models/joomla_component.php
@@ -29,6 +29,7 @@ use VDM\Joomla\Utilities\ObjectHelper;
 use VDM\Joomla\Utilities\GuidHelper;
 use VDM\Joomla\FOF\Encrypt\AES;
 use VDM\Joomla\Utilities\ArrayHelper as UtilitiesArrayHelper;
+use VDM\Joomla\Utilities\String\ComponentCodeNameHelper;
 use VDM\Joomla\Utilities\GetHelper;
 use VDM\Joomla\Componentbuilder\Extrusion\Helper\Extrusion;
 
@@ -1407,6 +1408,9 @@ class ComponentbuilderModelJoomla_component extends AdminModel
 		{
 			$data['system_name'] = $data['name'];
 		}
+
+		// make sure that the component code name is safe.
+		$data['name_code'] = ComponentCodeNameHelper::safe($data['name_code']);
 
 		// Set the GUID if empty or not valid
 		if (empty($data['guid']) && $data['id'] > 0)